mkosi: fix git_files_ignore()
Previously all directories would always be included in the copy
operation, which defeated half the point of the .gitignore logic, as
listing directories in it had no effect.

This resulted in an endless recursive copy operations of source trees if
they included a .mkosi- directory mkosi itself was outputting to.